@@ -43,8 +43,13 @@
     "cucumber": "^2.0.0-rc.9",
     "enzyme": "^2.6.0",
     "enzyme-to-json": "^1.5.1",
-    "eslint": "^3.8.1",
-    "eslint-plugin-react": "^6.5.0",
+    "eslint": "^3.19.0",
+    "eslint-config-airbnb": "^15.0.1",
+    "eslint-plugin-import": "^2.3.0",
+    "eslint-plugin-jest": "^20.0.3",
+    "eslint-plugin-jsx-a11y": "^5.0.3",
+    "eslint-plugin-react": "^7.0.1",
+    "eslint-watch": "^3.1.0",
     "extract-text-webpack-plugin": "^1.0.1",
     "file-loader": "^0.10.1",
     "ignore-styles": "^5.0.1",
@@ -70,7 +75,6 @@
     "body-parser": "^1.17.1",
     "compression": "^1.6.2",
     "cors": "^2.8.3",
-    "eslint-plugin-jest": "^20.0.3",
     "express": "^4.14.0",
     "express-session": "^1.15.3",
     "express-validator": "^3.2.0",
