@@ -2,29 +2,29 @@ import React from 'react';
 import styled from 'styled-components';
 
 export class FunctionButton extends React.PureComponent {
-	static propTypes = {
-		func: React.PropTypes.func,
-		text: React.PropTypes.string,
-		disabled: React.PropTypes.bool,
-	};
+  static propTypes = {
+    func: React.PropTypes.func,
+    text: React.PropTypes.string,
+    disabled: React.PropTypes.bool,
+  };
 
-	static defaultProps = {
-		func: () => null,
-	};
+  static defaultProps = {
+    func: () => null,
+  };
 
-	render() {
-		const { className, text, disabled } = this.props;
-		const { func } = this.props;
-		return (
-			<button
-				className={className}
-				onClick={() => func()}
-				disabled={disabled}
-			>
-				{text}
-			</button>
-		);
-	}
+  render() {
+    const { className, text, disabled } = this.props;
+    const { func } = this.props;
+    return (
+      <button
+        className={className}
+        onClick={() => func()}
+        disabled={disabled}
+      >
+        {text}
+      </button>
+    );
+  }
 }
 
 export default styled(FunctionButton)`
