@@ -4,8 +4,8 @@ import hangman from './hangman';
 import { routerReducer as routing } from 'react-router-redux';
 
 const reducer = combineReducers({
-	hangman,
-	routing,
+  hangman,
+  routing,
 });
 
 export default reducer;
