@@ -1,6 +1,6 @@
-export default socket => store => next => action => {
-	if (action.remote){
-		socket.emit('action', action);
-	}
-	return next(action);
+export default socket => store => next => (action) => {
+  if (action.remote) {
+    socket.emit('action', action);
+  }
+  return next(action);
 };
