@@ -5,10 +5,10 @@ import Hangman from './containers/hangman';
 import Management from './containers/management';
 
 const routes = (
-	<Route>
-		<Route path="/" component={Hangman} />
-		<Route path="/management" component={Management} />
-	</Route>
+  <Route>
+    <Route path="/" component={Hangman} />
+    <Route path="/management" component={Management} />
+  </Route>
 );
 
 export default routes;
