@@ -1,6 +1,6 @@
-import {createStore} from 'redux';
+import { createStore } from 'redux';
 import reducer from 'server/reducers/hangman';
 
-export default function setupStore(){
-	return createStore(reducer);
+export default function setupStore() {
+  return createStore(reducer);
 }
