@@ -1,14 +1,14 @@
 export default {
-	beforeEach: browser => {
-		browser
+  beforeEach: (browser) => {
+    browser
 			.url('http://localhost:3000')
 			.waitForElementVisible('body')
 			.waitForElementVisible('#app > div');
-	},
-	'Smoke test': browser => {
-		browser
+  },
+  'Smoke test': (browser) => {
+    browser
 			.assert.visible('#app > div', 'If app has rendered with React')
 			.assert.title('Hangman');
-	},
-	after: browser => browser.end(),
+  },
+  after: browser => browser.end(),
 };
